requirements.txt
pyqt5~=5.15.9
scikit-learn~=1.5.2
python-docx~=1.1.2
pywin32
pdfminer-six